audit: eliminate per-iteration GPU allocs, checked dispatch casts, DRY region SSOT
WGPU region.rs:
- Rewrite: reusable variants are the canonical SSOT implementation; dead
  allocating wrappers removed (no callers remain)
- checked_wg_x() enforces u32::try_from on all dispatch_workgroups calls
- region_u32() + region_meta() helpers collapse 5x repeated u32-conversion
  patterns into a single call site each

WGPU lu_decompose_blocked:
- Pre-allocate col_compact_buf (n * block_size) and row_compact_buf
  (block_size * n) before the panel loop; all 4 region transfer calls now
  use the reusable variants — zero per-iteration GPU allocations

WGPU cholesky_decompose_blocked:
- Pre-allocate panel_compact_buf (n * block_size) before the panel loop;
  all 3 region transfer calls use the reusable variant

WGPU linalg.rs:
- matmul_into and batched_matmul_into: dispatch dimensions use checked
  u32::try_from — eliminates silent truncation on large matrices

CUDA region.rs:
- Both device byte offset calculations use checked_mul + u64::try_from —
  eliminates silent overflow on large GPU buffers

+++ crates/hephaestus-cuda/src/application/sparse/spmm.rs
@@ -155,7 +155,7 @@ pub fn spmm_into<'a, T: CudaScalar + leto_ops::Scalar + Pod, B: AsGpuMatrixOpera
                 grid,
                 1,
                 1,
-                width.get() as u32,
+                width.get(),
                 1,
                 1,
                 0,
